Refactoring the logic to only take care of searching for cycles in the BOM and not computing the lowest level
