+++ backend/de.metas.manufacturing/src/main/java/org/eevolution/model/validator/PP_Product_BOMLine.java
@@ -37,9 +37,11 @@ import org.compiere.model.ModelValidator;
 import org.eevolution.api.BOMComponentType;
 import org.eevolution.api.IProductBOMBL;
 import org.eevolution.api.IProductBOMDAO;
+import org.eevolution.exceptions.BOMCycleException;
 import org.eevolution.model.I_PP_Order_BOMLine;
 import org.eevolution.model.I_PP_Product_BOMLine;
 
+import javax.swing.tree.DefaultMutableTreeNode;
 import java.util.List;
 
 @Validator(I_PP_Product_BOMLine.class)
@@ -127,16 +129,19 @@ public class PP_Product_BOMLine
 		return true;
 	}
 
-	@ModelChange(timings = { ModelValidator.TYPE_AFTER_NEW, ModelValidator.TYPE_AFTER_CHANGE, ModelValidator.TYPE_AFTER_DELETE }, skipIfCopying=true)
-	public void updateProductLowestLevelCode(final I_PP_Product_BOMLine bomLine)
+	@ModelChange(timings = { ModelValidator.TYPE_AFTER_NEW, ModelValidator.TYPE_AFTER_CHANGE }, skipIfCopying = true)
+	public void checkingBOMCycle(final I_PP_Product_BOMLine bomLine)
 	{
 		final ProductId productId = ProductId.ofRepoId(bomLine.getM_Product_ID());
 		final I_M_Product product = Services.get(IProductBL.class).getById(productId);
-		final int lowLevel = Services.get(IProductBOMBL.class).calculateProductLowestLevel(ProductId.ofRepoId(product.getM_Product_ID()));
+		try
+		{
+			Services.get(IProductBOMBL.class).createParentProductNodeForBOMLine(bomLine);
+		}
+		catch (final BOMCycleException e)
+		{
+			throw new LiberoException("Cycle detected in BOM for product: " + product.getValue());
+		}
 
-		product.setLowLevel(lowLevel); // update lowlevel
-		InterfaceWrapperHelper.save(product);
-
-		// TODO: if product's low level changed, we need to update the low level from all bom components where this product is in bom header.
 	}
 }
